Let talkback read People app name on launch
The root cause was that we were effectively doing setTitle(null) in onCreate.

We were replacing the people activity title with an intent extra, even if it's
not set.  Let's just remove this logic entirely because there's no point of
doing this, as the People activity doesn't show the app title.

This intent extra is not public API in ContactsContract, but it was in the
legacy API, so let's still keep ContactsRequest.set/getActivityTitle for now.
It still works for picker activities.

Bug 5508368

Change-Id: Id0ef790f9a7a721682de19d2c26d25e3622cee2d